+++ src/main/java/com/cheocharm/MapZ/usergroup/domain/repository/UserGroupRepository.java
@@ -1,8 +1,8 @@
-package com.cheocharm.MapZ.usergroup.repository;
+package com.cheocharm.MapZ.usergroup.domain.repository;
 
 import com.cheocharm.MapZ.group.domain.GroupEntity;
 import com.cheocharm.MapZ.user.domain.UserEntity;
-import com.cheocharm.MapZ.usergroup.UserGroupEntity;
+import com.cheocharm.MapZ.usergroup.domain.UserGroupEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
