+++ src/test/java/com/cheocharm/MapZ/usergroup/repository/UserGroupRepositoryCustomImplTest.java
@@ -6,11 +6,12 @@ import com.cheocharm.MapZ.group.domain.repository.GroupRepository;
 import com.cheocharm.MapZ.user.domain.UserEntity;
 import com.cheocharm.MapZ.user.domain.UserProvider;
 import com.cheocharm.MapZ.user.domain.repository.UserRepository;
-import com.cheocharm.MapZ.usergroup.InvitationStatus;
-import com.cheocharm.MapZ.usergroup.UserGroupEntity;
-import com.cheocharm.MapZ.usergroup.UserRole;
-import com.cheocharm.MapZ.usergroup.repository.vo.ChiefUserImageVO;
-import com.cheocharm.MapZ.usergroup.repository.vo.CountUserGroupVO;
+import com.cheocharm.MapZ.usergroup.domain.InvitationStatus;
+import com.cheocharm.MapZ.usergroup.domain.UserGroupEntity;
+import com.cheocharm.MapZ.usergroup.domain.UserRole;
+import com.cheocharm.MapZ.usergroup.domain.repository.UserGroupRepository;
+import com.cheocharm.MapZ.usergroup.domain.repository.vo.ChiefUserImageVO;
+import com.cheocharm.MapZ.usergroup.domain.repository.vo.CountUserGroupVO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
